# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Årsredovisningar
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of annual reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnnualReport"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a new annual report
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnualReport"
  /reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report by ID
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Annual report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnualReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update report metadata
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: Report updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnualReport"
  /reports/{reportId}/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get report completion summary
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report completion summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
...
  /reports/{reportId}/snapshots:
    get:
      operationId: listProjectSnapshots
      tags: [snapshots]
      summary: List point-in-time snapshots for a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Snapshot list
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items: { $ref: "#/components/schemas/ProjectSnapshot" }
                required: [snapshots]
    post:
      operationId: createProjectSnapshot
      tags: [snapshots]
      summary: Create a named snapshot of the current report state
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string }
              required: [label]
      responses:
        "201":
          description: Created snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSnapshot"

  # ─── Audit feed ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    AnnualReport:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        companyName:
          type: string
        fiscalYearStart:
          type: string
        fiscalYearEnd:
          type: string
        status:
          type: string
          enum: [draft, in_progress, complete, exported]
        accountingFramework:
          type: string
          enum: [K2, K3]
        completionPercent:
          type: integer
          description: Overall completion 0-100
        sectionsCompleted:
          type: integer
        sectionsTotal:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - companyName
        - fiscalYearStart
        - fiscalYearEnd
        - status
        - accountingFramework
        - completionPercent
        - sectionsCompleted
        - sectionsTotal
        - createdAt
        - updatedAt
    CreateReportBody:
      type: object
      properties:
        fiscalYearStart:
          type: string
        fiscalYearEnd:
          type: string
        accountingFramework:
          type: string
          enum: [K2, K3]
      required:
        - fiscalYearStart
        - fiscalYearEnd
        - accountingFramework
    UpdateReportBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, in_progress, complete, exported]
    ReportSummary:
      type: object
      properties:
        reportId:
          type: string
        companyName:
          type: string
        fiscalYearStart:
          type: string
        fiscalYearEnd:
          type: string
        status:
          type: string
          enum: [draft, in_progress, complete, exported]
        completionPercent:
          type: integer
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionStatus"
      required:
        - reportId
        - companyName
        - fiscalYearStart
        - fiscalYearEnd
        - status
        - completionPercent
        - sections
    SectionStatus:
      type: object
      properties:
        key:
          type: string
          description: "Section identifier (e.g. balance_sheet, income_statement, notes)"
        label:
          type: string
          description: Human-readable section name in Swedish
        completed:
          type: boolean
        requiredFields:
          type: integer
        completedFields:
          type: integer
      required:
        - key
        - label
        - completed
        - requiredFields
        - completedFields
...
    ProjectSnapshot:
      type: object
      properties:
        id: { type: string, format: uuid }
        reportId: { type: string, format: uuid }
        label: { type: string, nullable: true }
        actorProfileId: { type: string, format: uuid, nullable: true }
        actorName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, reportId, createdAt]
...